refactor: extract board alerts module
Forward-reference pattern again: getShareState / toggleShare still live in
server.mjs (moving to lib/share.mjs next); alerts.mjs imports them
circularly from ../server.mjs, safe because calls only happen inside
startAlertCron's setInterval callbacks, never at module-evaluation time.
